CO-2310 - Support optional detail rows
By default detail rows are hidden, with only the option toggle them one by one or globally. This provides a support for a rowDetailFn which is executed for each row and is decides whether we should render the row. For example, in a table containing days of the week, we could check if row.day === 'tuesday' and just show a detail row for that day.